% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{compute_jacard}
\alias{compute_jacard}
\title{Calculates Jaccard index}
\usage{
compute_jacard(poly1, poly2)
}
\value{
a numeric
}
\description{
The Jaccard index betwen two polygons is the proportion of the area
of intersection over the area of union.
}
