% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{gen_crosswalk}
\alias{gen_crosswalk}
\title{Generate crosswalk table}
\usage{
gen_crosswalk(pepfar, census)
}
\arguments{
\item{pepfar}{a PEPFAR sf object for a country at a given ADM level}

\item{census}{a Census sf object for a country at a given ADM level}
}
\value{
a df with a final crosswalk, with pepfar polygons as the
basis of comparison
}
\description{
This is a wrapper function that takes two sets of polygons
and generates a table of one-to-one matches, many-to-1, and partial matches.
}
