% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{merge_contained_pairs}
\alias{merge_contained_pairs}
\title{Assigns an existing crosswalk ID to polygons in \code{contained_df}.}
\usage{
merge_contained_pairs(one_to_one, many_to_one, contained_df)
}
\value{
a df
}
\description{
(INTERNAL)
}
