% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{process_1to1}
\alias{process_1to1}
\title{Find one-to-one matches between two sets of sf polygons}
\usage{
process_1to1(pepfar, census)
}
\value{
a df
}
\description{
For all polygons in the pepfar polygon set, finds all one-to-one
matches by computing a Jaccard index (INTERNAL). Rows with a Jaccard index above
the \code{get_option("one_to_one")} threshold are kept and returned.
}
