% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{process_contained_pairs}
\alias{process_contained_pairs}
\title{Find contained pairs between two sets of sf polygons}
\usage{
process_contained_pairs(pepfar, census)
}
\value{
a list of three dfs and one vectors
}
\description{
Looks for the census polygons that are contained by pepfar polygons
and returns a df, \code{contained_df}, where each row is a census poly contained by
a pepfar poly. If a census poly takes up more than 50\% of the area of a pepfar
poly, it is filtered out.
}
