% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{process_crosswalk}
\alias{process_crosswalk}
\title{Creates the final crosswalk table.}
\usage{
process_crosswalk(
  one_to_one,
  many_to_one,
  contained_summary,
  pepfar_names,
  census_names
)
}
\value{
a df
}
\description{
\code{one_to_one} is first flattened to match the dimensions of
\code{many_to_one}. \code{one_to_one} and \code{many_to_one} are then combined (row-binded)
and several new columns are created to record pertinent crosswalk information.
}
