% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{process_data}
\alias{process_data}
\title{Initial sf data processing for crosswalk generation}
\usage{
process_data(pepfar, census)
}
\value{
a list of dfs
}
\description{
(INTERNAL) Processes two sets of mutually comparable sf objects
for crosswalk comparison. This is the first step of the crosswalking heuristic.
}
