% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_crosswalk.R
\name{process_manyto1}
\alias{process_manyto1}
\title{Find many-to-one matches between two sets of sf polygons}
\usage{
process_manyto1(one_to_one, contained_census_ids, pepfar, census)
}
\value{
a df
}
\description{
For all polygons in the pepfar polygon set, finds the census polygons
that comprise them. Uses the thresholds, \code{get_option("many_to_one")} and
\code{get_option("census_contribution")}, to determine which polygons should be
included. (INTERNAL)
}
