% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sng_set}
\alias{sng_set}
\title{Set sugnatgeoalignR options}
\usage{
sng_set(
  one_to_one = NULL,
  many_to_one = NULL,
  census_contibution = NULL,
  buffer_dist = NULL
)
}
\description{
Set sugnatgeoalignR options
}
